requirements.txt
# Core RunPod dependency
runpod>=1.6.0

# Diffusers and Transformers
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# Image processing
Pillow>=10.0.0

# Other core dependencies
numpy>=1.24.0
requests>=2.28.0

# HuggingFace hub
huggingface_hub>=0.19.0

# PEFT for LoRA support
peft>=0.7.0

# SafeTensors for model loading
safetensors>=0.4.0

# Optional but recommended for performance
xformers>=0.0.20 